requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.7.0
django-ckeditor-5==0.2.18
django-cloudinary-storage==0.3.0
cloudinary==1.44.0
dj-database-url==2.3.0
gunicorn==23.0.0
whitenoise==6.9.0
psycopg2-binary==2.9.10
pillow==11.1.0
beautifulsoup4
python-decouple